requirements.txt
numpy>=1.24
scipy>=1.10
pydantic>=2.0
pyyaml>=6.0
scikit-image>=0.21
pytest>=7.0
